move top-level components to their own directory
