+++ src/components/Main/AppWithForm.jsx
@@ -1,0 +1,15 @@
+import React from 'react'
+import App from './App'
+import { Form } from '../../views'
+
+class AppWithForm extends React.Component {
+  render () {
+    return (
+      <App {...this.props}>
+        <Form {...this.props} />
+      </App>
+    )
+  }
+}
+
+export default AppWithForm
