+++ src/components/Main/Main.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+class Main extends React.Component {
+  render () {
+    return this.props.children
+  }
+}
+
+export default Main
